+++ lib/now-and-timers.cjs
