+++ lib/runner.js
@@ -7,6 +7,7 @@ const parseTestArgs = require('./parse-test-args');
 const snapshotManager = require('./snapshot-manager');
 const serializeError = require('./serialize-error');
 const Runnable = require('./test');
+const {waitForReady} = require('./worker/state.cjs');
 
 class Runner extends Emittery {
 	constructor(options = {}) {
@@ -53,7 +54,7 @@ class Runner extends Emittery {
 			serial: [],
 			todo: []
 		};
-		this.waitForReady = [];
+		this.waitForReady = waitForReady;
 
 		const uniqueTestTitles = new Set();
 		this.registerUniqueTitle = title => {
