+++ lib/test.js
@@ -3,7 +3,7 @@ const concordance = require('concordance');
 const isPromise = require('is-promise');
 const plur = require('plur');
 const assert = require('./assert');
-const nowAndTimers = require('./now-and-timers');
+const nowAndTimers = require('./now-and-timers.cjs');
 const parseTestArgs = require('./parse-test-args');
 const concordanceOptions = require('./concordance-options').default;
