+++ lib/worker/base.js
@@ -1,28 +1,12 @@
 'use strict';
 const {pathToFileURL} = require('url');
-const path = require('path');
 const currentlyUnhandled = require('currently-unhandled')();
-const {isRunningInThread, isRunningInChildProcess} = require('./utils');
+const {isRunningInThread, isRunningInChildProcess} = require('./utils.cjs');
 
-// Check if the test is being run without AVA cli
-if (!isRunningInChildProcess && !isRunningInThread) {
-	const chalk = require('chalk'); // Use default Chalk instance.
-	if (process.argv[1]) {
-		const fp = path.relative('.', process.argv[1]);
-
-		console.log();
-		console.error(`Test files must be run with the AVA CLI:\n\n    ${chalk.grey.dim('$')} ${chalk.cyan('ava ' + fp)}\n`);
-
-		process.exit(1);
-	} else {
-		throw new Error('The ’ava’ module can only be imported in test files');
-	}
-}
-
-const channel = require('./channel');
+const channel = require('./channel.cjs');
 
 const run = async options => {
-	require('./options').set(options);
+	require('./options.cjs').set(options);
 	require('../chalk').set(options.chalkOptions);
 
 	if (options.chalkOptions.level > 0) {
@@ -30,14 +14,14 @@ const run = async options => {
 		global.console = Object.assign(global.console, new console.Console({stdout, stderr, colorMode: true}));
 	}
 
-	const nowAndTimers = require('../now-and-timers');
+	const nowAndTimers = require('../now-and-timers.cjs');
 	const providerManager = require('../provider-manager');
 	const Runner = require('../runner');
 	const serializeError = require('../serialize-error');
 	const dependencyTracking = require('./dependency-tracker');
 	const lineNumberSelection = require('./line-numbers');
 
-	const sharedWorkerTeardowns = [];
+	const {sharedWorkerTeardowns} = require('./state.cjs');
 
 	async function exit(code) {
 		if (!process.exitCode) {
@@ -138,19 +122,6 @@ const run = async options => {
 	exports.getRunner = () => {
 		accessedRunner = true;
 		return runner;
-	};
-
-	exports.registerSharedWorker = (filename, initialData, teardown) => {
-		const {channel: sharedWorkerChannel, forceUnref, ready} = channel.registerSharedWorker(filename, initialData);
-		runner.waitForReady.push(ready);
-		sharedWorkerTeardowns.push(async () => {
-			try {
-				await teardown();
-			} finally {
-				forceUnref();
-			}
-		});
-		return sharedWorkerChannel;
 	};
 
 	// Store value to prevent required modules from modifying it.
