+++ lib/worker/channel.cjs
@@ -4,9 +4,9 @@ const {MessageChannel} = require('worker_threads');
 
 const pEvent = require('p-event');
 
-const timers = require('../now-and-timers');
-const {get: getOptions} = require('./options');
-const {isRunningInChildProcess, isRunningInThread} = require('./utils');
+const timers = require('../now-and-timers.cjs');
+const {get: getOptions} = require('./options.cjs');
+const {isRunningInChildProcess, isRunningInThread} = require('./utils.cjs');
 
 const selectAvaMessage = type => message => message.ava && message.ava.type === type;
