+++ lib/worker/dependency-tracker.js
@@ -1,6 +1,6 @@
 /* eslint-disable node/no-deprecated-api */
 'use strict';
-const channel = require('./channel');
+const channel = require('./channel.cjs');
 
 const seenDependencies = new Set();
 let newDependencies = [];
