+++ lib/worker/main.js
@@ -1,3 +1,5 @@
 'use strict';
+require('./guard-environment.cjs'); // eslint-disable-line import/no-unassigned-import
+
 const runner = require('./base').getRunner();
 module.exports = runner.chain;
