+++ lib/worker/options.cjs
