+++ lib/worker/plugin.cjs
@@ -1,12 +1,23 @@
 const pkg = require('../../package.json');
-const baseWorker = require('./base');
-const options = require('./options');
+const {registerSharedWorker: register} = require('./channel.cjs');
+const options = require('./options.cjs');
+const {sharedWorkerTeardowns, waitForReady} = require('./state.cjs');
+
+require('./guard-environment.cjs'); // eslint-disable-line import/no-unassigned-import
 
 const workers = new Map();
 const workerTeardownFns = new WeakMap();
 
 function createSharedWorker(filename, initialData, teardown) {
-	const channel = baseWorker.registerSharedWorker(filename, initialData, teardown);
+	const {channel, forceUnref, ready} = register(filename, initialData, teardown);
+	waitForReady.push(ready);
+	sharedWorkerTeardowns.push(async () => {
+		try {
+			await teardown();
+		} finally {
+			forceUnref();
+		}
+	});
 
 	class ReceivedMessage {
 		constructor(id, data) {
