+++ lib/worker/state.cjs
@@ -1,0 +1,3 @@
+'use strict';
+exports.sharedWorkerTeardowns = [];
+exports.waitForReady = [];
