+++ lib/worker/utils.cjs
